+++ completions/bashunit.bash
@@ -17,14 +17,14 @@ _BASHUNIT_COMPLETIONS_DOC_OPTS="--custom -e --env --boot -h --help"
 _BASHUNIT_COMPLETIONS_TEST_OPTS="--assert --boot --changed --coverage --coverage-exclude \
 --coverage-diff --coverage-min --coverage-paths --coverage-report --coverage-report-cobertura --coverage-report-html \
 --debug --detailed --dry-run --env --exclude-filter --exclude-tag --fail-on-flaky --fail-on-risky --failures-only \
---filter --gha-annotations --help --jobs --list --list-format --log-gha --log-junit --login --no-color \
+--filter --gha-annotations --help --jobs --list --list-format --list-suites --log-gha --log-junit --login --no-color \
 --no-coverage-report --no-output --no-output-on-failure --no-parallel \
 --no-progress --no-snapshot-create --order-by --output --parallel --profile \
 --random-order --repeat --report-html \
 --report-json --report-junit --report-md --report-tap --rerun-failed --retry --run-all \
 --seed --shard --show-incomplete --show-output --show-skipped --simple \
 --skip-env-file --snapshot-report-unused --snapshot-update \
---stop-on-failure --strict --tag \
+--stop-on-failure --strict --suite --tag \
 --test-timeout --verbose \
 --watch -R -S -a -e -f -h -j -l -p -r -s -vvv -w"
 
@@ -88,7 +88,7 @@ _bashunit_completions() {
     COMPREPLY=($(compgen -W "auto always never" -- "$cur"))
     return 0
     ;;
-  -f | --filter | --exclude-filter | --tag | --exclude-tag | --repeat | --retry | --seed | --shard | --test-timeout)
+  -f | --filter | --exclude-filter | --suite | --tag | --exclude-tag | --repeat | --retry | --seed | --shard | --test-timeout)
     return 0
     ;;
   esac